+++ packages/@expo/cli/src/run/android/resolveGradlePropsAsync.ts
@@ -3,6 +3,9 @@ import path from 'path';
 import { Device, getDeviceABIsAsync } from '../../start/platforms/android/adb';
 import { CommandError } from '../../utils/errors';
 
+// Supported ABIs for Android. see https://developer.android.com/ndk/guides/abis
+const VALID_ARCHITECTURES = ['armeabi-v7a', 'arm64-v8a', 'x86', 'x86_64'];
+
 export type GradleProps = {
   /** Directory for the APK based on the `variant`. */
   apkVariantDirectory: string;
@@ -60,5 +63,7 @@ async function getConnectedDeviceABIS(
   }
 
   const abis = await getDeviceABIsAsync(device);
-  return abis.filter((abi, i, arr) => arr.indexOf(abi) === i).join(',');
+
+  const validAbis = abis.filter((abi) => VALID_ARCHITECTURES.includes(abi));
+  return validAbis.filter((abi, i, arr) => arr.indexOf(abi) === i).join(',');
 }